update: dependencies

+++ app-shared/src/database/tables/bug_message.rs
@@ -43,7 +43,7 @@ VALUES (DEFAULT, $1, $2);
 ",
         )
         .bind(author_id.0 as i64)
-        .bind(issue_number)
+        .bind(issue_number as i64)
         .execute(pool)
         .await
     }
